+++ silvahnos-be/src/main/java/com/app/silvahnosbe/entities/FacturaEntity.java
@@ -30,8 +30,8 @@ public class FacturaEntity {
     private Integer monto;
     private String observaciones;
     private boolean borrado;
-    private int empresa;
-    private int estado;
+    private Long empresa;
+    private Long estado;
 
     // Fechas de acciones
     @CreationTimestamp
